Remove build script from Cargo.toml
The build script is only required for the tests. However those tests
aren't included when packaging the crate, which cause the build script
to fail.

By removing the build field from Cargo.toml allows the package command
to succeed, without include any tests or test data. But because we use
the default convention for the build script we can still use it when we
have the entire repo (tests and test data included).